@@ -5,8 +5,8 @@ homepage: https://github.com/fluttercandies/pull_to_refresh_notification
 
 
 environment:
-  sdk: '>=2.12.0 <3.0.0'
-  flutter: ">=2.0.0"
+  sdk: '>=3.0.0 <4.0.0'
+  flutter: ">=3.0.0"
 
 dependencies:
   flutter:
